Move global interaction zone tracing into the correct spot.
